+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+- Buttons StreamBlock for WhatsApp messages
+
+### Changed
+- Migrated next_prompt on WhatsApp messages to buttons
+- Removed next_prompt on WhatsApp messages
+
 ## v1.1.0-dev.3
 
 ### Fixed